requirements.txt
crewai==0.70.1
crewai[tools]==0.70.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
markdown==3.5.1

